+++ MoMoney/App.xaml.cs
@@ -8,7 +8,6 @@ public partial class App : Application
 {
     public App()
     {
-        Syncfusion.Licensing.SyncfusionLicenseProvider.RegisterLicense(Secret.SfLicenseKey);
         InitializeComponent();
     }