+++ MoMoney/Views/Settings/BulkEditingPage.xaml.cs
@@ -11,25 +11,6 @@ public partial class BulkEditingPage : ContentPage
 		InitializeComponent();
 		vm = _vm;
 		BindingContext = vm;
-
-		PckFindAccount.SelectedValueChanged += OnClear;
-		PckFindCategory.SelectedValueChanged += OnClear;
-		FindPayee.SelectionChanged += OnClear;
-
-		PckReplaceAccount.SelectedValueChanged += OnClear;
-        PckReplaceCategory.SelectedValueChanged += OnClear;
-		ReplacePayee.TextChanged += OnClear;
-
-		PckFindCategory.SelectedValueChanged += vm.GetFindSubcategories;
-		PckReplaceCategory.SelectedValueChanged += vm.GetReplaceSubcategories;
-	}
-
-	void OnClear(object? sender, object newValue)
-	{
-		if (newValue == null)
-		{
-			vm.Clear();
-        }
 	}
 
     protected override async void OnAppearing()